+++ CHANGELOG.rst
@@ -15,6 +15,7 @@ Changed
 
 * vimrc to use vim-plug instead of Vundle
 * Only install .taskrc if taskwarrior is present
+* only add bash completion for pipenv if pipenv is present
 
 0.1.0 - 2018-08-05
 ==================